refactor: Extract createClient function from client.js
